+++ app/Livewire/Dashboard/UsuariosComponent.php
@@ -2,7 +2,6 @@
 
 namespace App\Livewire\Dashboard;
 
-use App\Models\Empresa;
 use App\Models\Parametro;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
@@ -11,32 +10,53 @@ use Illuminate\Validation\Rule;
 use Jantinnerezo\LivewireAlert\LivewireAlert;
 use Livewire\Attributes\On;
 use Livewire\Component;
-use Livewire\WithPagination;
 
 class UsuariosComponent extends Component
 {
     use LivewireAlert;
-    use WithPagination;
 
+    public $rows = 0, $numero = 14, $tableStyle = false;
     public $view = "create", $keyword;
     public $name, $email, $password, $role, $usuarios_id;
     public $edit_name, $edit_email, $edit_password, $edit_role = 0, $edit_roles_id = 0, $created_at, $estatus = 1, $photo, $empresas_id;
     public $rol_nombre, $tabla = 'usuarios', $getPermisos, $cambios = false;
-    public $select_empresas, $ver_empresas; // acceso a empresas
+
+    public function mount()
+    {
+        $this->setLimit();
+    }
 
     public function render()
     {
         $roles = Parametro::where('tabla_id', '-1')->get();
+
         $users = User::buscar($this->keyword)
             ->orderBy('role', 'DESC')
             ->orderBy('roles_id', 'DESC')
             ->orderBy('created_at', 'DESC')
-            ->paginate(numRowsPaginate());
+            ->limit($this->rows)
+            ->get();
+
         $rows = User::count();
+
+        if ($rows > $this->numero) {
+            $this->tableStyle = true;
+        }
+
         return view('livewire.dashboard.usuarios-component')
             ->with('listarRoles', $roles)
             ->with('listarUsers', $users)
-            ->with('rows', $rows);
+            ->with('rowsUsuarios', $rows);
+    }
+
+    public function setLimit()
+    {
+        if (numRowsPaginate() < $this->numero) {
+            $rows = $this->numero;
+        } else {
+            $rows = numRowsPaginate();
+        }
+        $this->rows = $this->rows + $rows;
     }
 
     #[On('limpiar')]
@@ -47,6 +67,7 @@ class UsuariosComponent extends Component
             'edit_name', 'edit_email', 'edit_password', 'edit_role', 'edit_roles_id', 'created_at', 'estatus',
             'photo', 'empresas_id', 'rol_nombre', 'getPermisos', 'cambios'
         ]);
+        $this->resetErrorBag();
     }
 
     public function generarClave()
@@ -85,8 +106,6 @@ class UsuariosComponent extends Component
     {
         $this->validate($this->rules($this->usuarios_id));
 
-        $this->accesoEmpresa($this->select_empresas); // acceso a empresas
-
         if (is_null($this->usuarios_id)) {
             //nuevo
             $usuarios = new User();
@@ -110,101 +129,94 @@ class UsuariosComponent extends Component
         } else {
             //editar
             $usuarios = User::find($this->usuarios_id);
-            $usuarios->name = ucwords($this->edit_name);
-            $usuarios->email = strtolower($this->edit_email);
-            if ($this->edit_role > 1) {
-                $usuarios->role = 2;
-                $usuarios->roles_id = $this->edit_role;
-                $role = Parametro::where('tabla_id', '-1')->where('id', $this->edit_role)->first();
-                if ($role) {
-                    $usuarios->permisos = $role->valor;
+            if ($usuarios){
+                $usuarios->name = ucwords($this->edit_name);
+                $usuarios->email = strtolower($this->edit_email);
+                if ($this->edit_role > 1) {
+                    $usuarios->role = 2;
+                    $usuarios->roles_id = $this->edit_role;
+                    $role = Parametro::where('tabla_id', '-1')->where('id', $this->edit_role)->first();
+                    if ($role) {
+                        $usuarios->permisos = $role->valor;
+                    }
+                } else {
+                    $usuarios->role = $this->edit_role;
+                    $usuarios->roles_id = null;
                 }
-            } else {
-                $usuarios->role = $this->edit_role;
-                $usuarios->roles_id = null;
+                $usuarios->save();
+                $this->edit($this->usuarios_id);
+                $this->alert('success', 'Usuario Actualizado.');
+            }else{
+                $this->dispatch('cerrarModal');
             }
-            $usuarios->update();
-            $this->edit($this->usuarios_id);
-            $this->alert('success', 'Usuario Actualizado.');
         }
     }
 
     public function edit($id)
     {
+        $this->limpiar();
         $usuario = User::find($id);
-        $this->usuarios_id = $usuario->id;
-        $this->edit_name = $usuario->name;
-        $this->edit_email = $usuario->email;
-        if ($usuario->roles_id) {
-            $this->edit_role = $usuario->roles_id;
-        }else{
-            $this->edit_role = $usuario->role;
-        }
-        $this->edit_roles_id = $usuario->roles_id;
-        $this->estatus = $usuario->estatus;
-        $this->created_at = $usuario->created_at;
-        $this->photo = $usuario->profile_photo_path;
-        $this->rol_nombre = verRole($usuario->role, $usuario->roles_id);
-        $this->getPermisos = $usuario->permisos;
-
-        //acceso a empresas
-
-        $getEmpresas = Empresa::orderBy('nombre', 'ASC')->get();
-        $data = array();
-        $placeholder = '';
-        foreach ($getEmpresas as $empresa){
-            $option = [
-                'id' => $empresa->id,
-                'text' => $empresa->nombre
-            ];
-            if (leerJson($empresa->permisos, $this->usuarios_id)){
-                $placeholder .= '['.$empresa->nombre.'] <br>';
+        if ($usuario){
+            $this->usuarios_id = $usuario->id;
+            $this->edit_name = $usuario->name;
+            $this->edit_email = $usuario->email;
+            if ($usuario->roles_id) {
+                $this->edit_role = $usuario->roles_id;
+            }else{
+                $this->edit_role = $usuario->role;
             }
-            array_push($data, $option);
+            $this->edit_roles_id = $usuario->roles_id;
+            $this->estatus = $usuario->estatus;
+            $this->created_at = $usuario->created_at;
+            $this->photo = $usuario->profile_photo_path;
+            /*$this->empresas_id = $usuario->empresas_id;*/
+            $this->rol_nombre = verRole($usuario->role, $usuario->roles_id);
+            $this->getPermisos = $usuario->permisos;
+        }else{
+            $this->dispatch('cerrarModal');
         }
-        $this->ver_empresas = $placeholder;
-        $this->dispatch('selectEmpresas', data: $data);
-
     }
 
     public function cambiarEstatus($id)
     {
         $usuario = User::find($id);
-        if ($usuario->estatus) {
-            $usuario->estatus = 0;
-            $texto = "Usuario Suspendido";
-        } else {
-            $usuario->estatus = 1;
-            $texto = "Usuario Activado";
+        if ($usuario){
+            if ($usuario->estatus) {
+                $usuario->estatus = 0;
+                $texto = "Usuario Suspendido";
+            } else {
+                $usuario->estatus = 1;
+                $texto = "Usuario Activado";
+            }
+            $usuario->update();
+            $this->estatus = $usuario->estatus;
+            $this->alert('success', $texto);
+        }else{
+            $this->dispatch('cerrarModal');
         }
-        $usuario->update();
-        $this->estatus = $usuario->estatus;
-        $this->alert(
-            'success',
-            $texto
-        );
     }
 
     public function restablecerClave($id)
     {
-        if (!$this->edit_password) {
-            $clave = Str::password(8);
-        } else {
-            $clave = $this->edit_password;
-        }
         $usuario = User::find($id);
-        $usuario->password = Hash::make($clave);
-        $usuario->update();
-        $this->edit_password = $clave;
-        $this->alert(
-            'success',
-            'Contraseña Restablecida.'
-        );
+        if ($usuario){
+            if (!$this->edit_password) {
+                $clave = Str::password(8);
+            } else {
+                $clave = $this->edit_password;
+            }
+            $usuario->password = Hash::make($clave);
+            $usuario->update();
+            $this->edit_password = $clave;
+            $this->alert('success', 'Contraseña Restablecida');
+        }else{
+            $this->dispatch('cerrarModal');
+        }
     }
 
     public function destroy($id)
     {
-        $this->usuarios_id = $id;
+        $this->usuario_id = $id;
         $this->confirm('¿Estas seguro?', [
             'toast' => false,
             'position' => 'center',
@@ -220,30 +232,32 @@ class UsuariosComponent extends Component
     public function confirmedUser()
     {
         $usuario = User::find($this->usuarios_id);
+        if ($usuario){
+            //codigo para verificar si realmente se puede borrar, dejar false si no se requiere validacion
+            $vinculado = false;
 
-        //codigo para verificar si realmente se puede borrar, dejar false si no se requiere validacion
-        $vinculado = false;
-
-        if ($vinculado) {
-            $this->alert('warning', '¡No se puede Borrar!', [
-                'position' => 'center',
-                'timer' => '',
-                'toast' => false,
-                'text' => 'El registro que intenta borrar ya se encuentra vinculado con otros procesos.',
-                'showConfirmButton' => true,
-                'onConfirmed' => '',
-                'confirmButtonText' => 'OK',
-            ]);
-        } else {
-            $usuario->delete();
-            $this->limpiar();
+            if ($vinculado) {
+                $this->alert('warning', '¡No se puede Borrar!', [
+                    'position' => 'center',
+                    'timer' => '',
+                    'toast' => false,
+                    'text' => 'El registro que intenta borrar ya se encuentra vinculado con otros procesos.',
+                    'showConfirmButton' => true,
+                    'onConfirmed' => '',
+                    'confirmButtonText' => 'OK',
+                ]);
+            } else {
+                $usuario->delete();
+                $this->limpiar();
+                $this->dispatch('cerrarModal');
+                $this->alert(
+                    'success',
+                    'Usuario Eliminado.'
+                );
+            }
+        }else{
             $this->dispatch('cerrarModal');
-            $this->alert(
-                'success',
-                'Usuario Eliminado.'
-            );
         }
-
     }
 
     #[On('buscar')]
@@ -278,10 +292,12 @@ class UsuariosComponent extends Component
 
     public function savePermisos(){
         $usuario = User::find($this->usuarios_id);
-        $usuario->permisos = $this->getPermisos;
-        $usuario->save();
-        $this->reset('cambios');
-        $this->alert('success', 'Permisos Guardados.');
+        if ($usuario){
+            $usuario->permisos = $this->getPermisos;
+            $usuario->save();
+            $this->reset('cambios');
+            $this->alert('success', 'Permisos Guardados.');
+        }
     }
 
     public function deletePermisos()
@@ -290,37 +306,21 @@ class UsuariosComponent extends Component
         $this->cambios = true;
     }
 
-    //acceso a empresas **********************************************************************
-
-    public function accesoEmpresa($array)
+    public function getEstatusUsuario($i, $icon = null)
     {
-        //acceso a aempresas
-        if (!empty($array)){
-            foreach ($array as $id){
-                $empresa = Empresa::find($id);
-                $permisos = json_decode($empresa->permisos, true);
-                if (!leerJson($empresa->permisos, $this->usuarios_id)){
-                    $permisos[$this->usuarios_id] = true;
-                }else{
-                    unset($permisos[$this->usuarios_id]);
-                }
-                $permisos = json_encode($permisos);
-                $empresa->permisos = $permisos;
-                $empresa->update();
-            }
+        if (is_null($icon)){
+            $suspendido = "Suspendido";
+            $activado = "Activo";
+        }else{
+            $suspendido = '<i class="fa fa-user-slash"></i>';
+            $activado = '<i class="fa fa-user-check"></i>';
         }
-    }
-
-    #[On('empresasSeleccionadas')]
-    public function empresasSeleccionadas($data)
-    {
-        $this->select_empresas = $data;
-    }
-
-    #[On('selectEmpresas')]
-    public function selectEmpresas($data)
-    {
-        //select acceso a empresas // JS
+        $status = [
+            '0' => '<span class="text-danger">'.$suspendido.'</span>',
+            '1' => '<span class="text-success">'.$activado.'</span>',
+            /*'2' => '<span class="text-success">Confirmado</span>'*/
+        ];
+        return $status[$i];
     }
 
 }
